requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.0
pytz==2023.3
schedule==1.2.0
